added user access limitations edited tests task update test checks if task exists, edits task, re-checks data

+++ Organizer/project/routes/tasks.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, request
-from flask_login.utils import login_required
+from flask_login.utils import login_required, current_user
 from project import db
 from project.models import Task
 
@@ -12,13 +12,13 @@ def get_tasks():
 
     """Reads the list of tasks"""
 
-    tasks = Task.query.all()
+    tasks = Task.query.filter_by(user_id = current_user.id)
 
     output = []
     for task in tasks:
         task_data = {
             "ToDo": task.task,
-            "Created": task.created_at,
+            "Created": task.created_at.strftime("%Y-%m-%d"),
             "status": task.completed,
             "id": task.id,
         }
@@ -32,25 +32,32 @@ def get_tasks():
 @login_required
 def add_task():
     """Creates a new task"""
-
-    task = Task(task=request.json["task"])
+    task = Task(task=request.json["task"], user_id = current_user.id)
 
     db.session.add(task)
     db.session.commit()
 
-    return {"id": task.id, "created": task.created_at}, 201
+    return {"id": task.id, "created": task.created_at.strftime("%Y-%m-%d")}, 201
 
+@tasks.route("/task/<id>", methods=["GET"])
+@login_required
+def get_task_byid(id):
+    task = Task.query.filter_by(user_id = current_user.id, id = id).first()
+
+    return {"id": task.id, "task": task.task, "status": task.completed}, 200
 
 @tasks.route("/task/<id>", methods=["PUT"])
 @login_required
 def edit_task(id):
     """Updates a task"""
 
-    task = Task.query.get_or_404(id)
+    task = Task.query.filter_by(user_id = current_user.id, id = id).first()
+    if not task:
+        return {"message": "Not found"}, 404
     task.task = request.json.get("task", task.task)
 
     db.session.commit()
-    return {"id": task.id, "edited task": task.task, "updated": task.updated_at}, 201
+    return {"id": task.id, "edited task": task.task, "updated": task.updated_at}, 202
 
 
 @tasks.route("/task/<id>", methods=["DELETE"])
@@ -58,7 +65,9 @@ def edit_task(id):
 def delete_task(id):
     """Deletes a task by id"""
 
-    task = Task.query.get_or_404(id)
+    task = Task.query.filter_by(user_id = current_user.id, id = id).first()
+    if not task:
+        return {"message": "Not found"}, 404
 
     db.session.delete(task)
     db.session.commit()
